- Split handlers from API in computational subsystem - Created mocker for computational subsistem in test_projects
